requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.4.0
audioop-lts==0.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
discord.py==2.6.3
dotenv==0.9.9
frozenlist==1.7.0
idna==3.10
propcache==0.4.0
multidict==6.7.0
pycparser==2.23
PyNaCl==1.6.0
python-dotenv==1.1.1
requests==2.32.5
urllib3==2.5.0
yarl==1.22.0
yt-dlp==2025.9.26
PyNaCl>=1.5.0
pytest>=8.4.2
mafic>=2.10.1